requirements.txt
streamlit==1.34.0
pandas==2.2.2
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.20.0
openpyxl==3.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
plotly==5.21.0